+++ app/src/main/java/com/techcourse/controller/ErrorController.java
@@ -1,4 +1,4 @@
-package com.techcourse.controllerV2;
+package com.techcourse.controller;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
