+++ app/src/main/java/com/techcourse/controller/RegisterController.java
@@ -4,18 +4,27 @@ import com.techcourse.domain.User;
 import com.techcourse.repository.InMemoryUserRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.controller.asis.Controller;
+import nextstep.mvc.view.ModelAndView;
+import nextstep.web.annotation.Controller;
+import nextstep.web.annotation.RequestMapping;
+import nextstep.web.support.RequestMethod;
 
-public class RegisterController implements Controller {
+@Controller
+public class RegisterController {
 
-    @Override
-    public String execute(final HttpServletRequest req, final HttpServletResponse res) throws Exception {
-        final var user = new User(2,
+    @RequestMapping(value = "/register/view", method = RequestMethod.GET)
+    public ModelAndView doGet(final HttpServletRequest req, final HttpServletResponse res) {
+        return new ModelAndView("/register.jsp");
+    }
+
+    @RequestMapping(value = "/register", method = RequestMethod.POST)
+    public ModelAndView doPost(final HttpServletRequest req, final HttpServletResponse res) {
+        User user = new User(2,
                 req.getParameter("account"),
                 req.getParameter("password"),
                 req.getParameter("email"));
         InMemoryUserRepository.save(user);
 
-        return "redirect:/index.jsp";
+        return new ModelAndView("redirect:/index.jsp");
     }
 }
