--- app/src/main/java/com/techcourse/controllerV2/LoginController.java
@@ -1,27 +1,0 @@
-package com.techcourse.controllerV2;
-
-import com.techcourse.controller.UserSession;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.view.ModelAndView;
-import nextstep.web.annotation.Controller;
-import nextstep.web.annotation.RequestMapping;
-import nextstep.web.support.RequestMethod;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-@Controller
-public class LoginController {
-
-    private static final Logger log = LoggerFactory.getLogger(LoginController.class);
-
-    @RequestMapping(value = "/login/view", method = RequestMethod.GET)
-    public ModelAndView doGet(final HttpServletRequest req, final HttpServletResponse res) {
-        return UserSession.getUserFrom(req.getSession())
-                .map(user -> {
-                    log.info("logged in {}", user.getAccount());
-                    return new ModelAndView("redirect:/index.jsp");
-                })
-                .orElse(new ModelAndView("/login.jsp"));
-    }
-}
